Update analysis.py
Richard, I updated the volcano function as much as I really could with my understanding of the data structures. I made a comment at the beginning stating the two things that really need to be done to make this a seamless integration into DATSIGMA. let me know if you need anything else. Sorry I wasn't able to completely incorporate it.